% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_hmm.R
\name{dhmm_bern}
\alias{dhmm_bern}
\title{-----------------------------------------------------------------------------
Binomial HMM distribution
-----------------------------------------------------------------------------}
\usage{
dhmm_bern(x, pi, PArray, GammaArray, len = 0, log = 0)
}
\description{
-----------------------------------------------------------------------------
Binomial HMM distribution
-----------------------------------------------------------------------------
}
