% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_hmm.R
\name{dhmm_binom}
\alias{dhmm_binom}
\title{-----------------------------------------------------------------------------
Binomial HMM distribution
-----------------------------------------------------------------------------}
\usage{
dhmm_binom(x, pi, PArray, size, GammaArray, len = 0, log = 0)
}
\description{
-----------------------------------------------------------------------------
Binomial HMM distribution
-----------------------------------------------------------------------------
}
