% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dJS_Do.R
\name{dJS}
\alias{dJS}
\title{-----------------------------------------------------------------------------
Bernoulli (standard)
-----------------------------------------------------------------------------}
\usage{
dJS(x, init, prob, probTrans, pstar = -1, len = 0, log = 0)
}
\description{
-----------------------------------------------------------------------------
Bernoulli (standard)
-----------------------------------------------------------------------------
}
