% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dJS_Do.R
\name{dJS_binom_Do}
\alias{dJS_binom_Do}
\title{-----------------------------------------------------------------------------
Binomial addition
-----------------------------------------------------------------------------}
\usage{
dJS_binom_Do(x, init, prob, size, probTrans, len, pstar = -1, log = 0)
}
\description{
-----------------------------------------------------------------------------
Binomial addition
-----------------------------------------------------------------------------
}
