% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dJS_Do.R
\name{dJS_Do}
\alias{dJS_Do}
\title{-----------------------------------------------------------------------------
-----------------------------------------------------------------------------}
\usage{
dJS_Do(x, init, probObs, probTrans, len, pstar = -1, checkRowSums = 1, log = 0)
}
\description{
-----------------------------------------------------------------------------
-----------------------------------------------------------------------------
}
