% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_dist.R
\name{dJS_ms}
\alias{dJS_ms}
\title{-----------------------------------------------------------------------------
Multistate
-----------------------------------------------------------------------------}
\usage{
dJS_ms(x, pstar, piVector, PArray, GammaArray, len, log = 0)
}
\description{
-----------------------------------------------------------------------------
Multistate
-----------------------------------------------------------------------------
}
