% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dJS_o.R
\name{dJS_o}
\alias{dJS_o}
\title{-----------------------------------------------------------------------------
-----------------------------------------------------------------------------}
\usage{
dJS_o(x, init, probObs, probTrans, len, pstar = -1, checkRowSums = 1, log = 0)
}
\description{
-----------------------------------------------------------------------------
-----------------------------------------------------------------------------
}
