% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JS_dists.R
\name{dJS_pois}
\alias{dJS_pois}
\title{-----------------------------------------------------------------------------
Poission observations
-----------------------------------------------------------------------------}
\usage{
dJS_pois(x, init, rate, probTrans, pstar, weight = 1, len = 0, log = 0)
}
\description{
-----------------------------------------------------------------------------
Poission observations
-----------------------------------------------------------------------------
}
