% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_dist.R
\name{dJS_rd}
\alias{dJS_rd}
\title{-----------------------------------------------------------------------------
Binomial robust design
-----------------------------------------------------------------------------}
\usage{
dJS_rd(x, pstar, piVector, PArray, nSubOcc, GammaArray, len = 0, log = 0)
}
\description{
-----------------------------------------------------------------------------
Binomial robust design
-----------------------------------------------------------------------------
}
