% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{make_nimble_icar}
\alias{make_nimble_icar}
\title{Create liest arguments for a random walk of order `p`.to be used
in the `dcar_normal` distribution}
\usage{
make_nimble_icar(n, p = 1)
}
\arguments{
\item{n}{The length of the RW(p) process. The length must be greater than or equal to \code{2*p}.}

\item{p}{The order of the random walk process.}
}
\value{
A list with elements: `adj`, `weights`, and `num` for use with the
 `dcar_normal` distribution in `nimble` models
}
\description{
This function calculates the entries for a precision matrix for a random walk of order `p`.
}
\references{
H. Rue and L. Held (2005) Gaussian Markov Random Fields. Chapman & Hall/CRC. 263 pp.
}
