% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JS_dists.R
\name{dJS_binom}
\alias{dJS_binom}
\title{-----------------------------------------------------------------------------
Binomial addition
-----------------------------------------------------------------------------}
\usage{
dJS_binom(x, init, prob, size, probTrans, pstar, weight = 1, len = 0, log = 0)
}
\description{
-----------------------------------------------------------------------------
Binomial addition
-----------------------------------------------------------------------------
}
